backend/gateway/appointment_service/requirements.txt
Flask
Flask-Cors
Flask-SQLAlchemy
python-dotenv
PyJWT